+++ test/react-router/codemods/fixtures/root/with-existing-sentry.tsx
@@ -1,6 +1,0 @@
-import * as Sentry from '@sentry/react-router';
-import { Outlet } from 'react-router';
-
-export default function App() {
-  return <Outlet />;
-}