+++ test/react-router/codemods/fixtures/root/with-isrouteerrorresponse.tsx
@@ -1,5 +1,0 @@
-import { Outlet, isRouteErrorResponse } from 'react-router';
-
-export default function App() {
-  return <Outlet />;
-}